+++ template/src/testclient/testchannels/$$channel$$.ts.js
